part2 2.15

+++ react-app/src/part 2/phonebook/components/DeleteBtn.js
@@ -3,11 +3,13 @@ import { getAll, deleteNumber} from '../backend/actions';
 
 const DeleteBtn = ({id, setPersons}) => {
     const handleOnClick = async (id) => {
-        const data = await deleteNumber(id)
+        if(window.confirm('Do you want to delete this number?')){
+            const data = await deleteNumber(id)
             .then(() => {
                 getAll()
                     .then(response => setPersons(response));
             })
+        }
     }
 
   return (
